build: Reetirado CRA e adicionado vite com TS
