--- ease-acesse/src/hooks/useLoading.ts
@@ -1,20 +1,0 @@
-import { useState } from "react";
-
-export function useLoading(){
-    const [ isLoading, setIsLoading ] = useState(false);
-
-
-    const startLoading = () => {
-        setIsLoading(true);
-    }
-
-    const stopLoading = () => {
-        setIsLoading(false);
-    }
-
-    return {
-        isLoading,
-        startLoading,
-        stopLoading,
-    };
-}
