--- ease-acesse/src/pages/Provisionamento/SearchONU/style.ts
@@ -1,14 +1,0 @@
-import styled from "styled-components";
-
-export const Form = styled.form`
-    width: 100%;
-    max-width: 500px;
-    flex-direction: column;
-    margin-bottom: 20px;
-
-    .MUI-CircularProgress{
-        width: 30px !important;
-        height: 30px !important;
-    }
-
-`;
