+++ src/interfaces/OltInfoItem.ts
